[refactor] mirror Backend folder structure for tests
