+++ Backend.Tests/Repository/UserRepositoryTests.cs
@@ -1,0 +1,6 @@
+namespace Backend.Tests.Repository;
+
+public class UserRepositoryTests
+{
+
+}
